Refactored common sample code into a shared project. and moved DependencyInjection.Cli project to consume it.
