+++ samples/D20Tek.Samples.Common/D20Tek.Samples.Common.projitems
@@ -1,0 +1,18 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <MSBuildAllProjects Condition="'$(MSBuildVersion)' == '' Or '$(MSBuildVersion)' &lt; '16.0'">$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
+    <HasSharedItems>true</HasSharedItems>
+    <SharedGUID>4f5494aa-3286-4e53-986d-a638d2bb2ccc</SharedGUID>
+  </PropertyGroup>
+  <PropertyGroup Label="Configuration">
+    <Import_RootNamespace>D20Tek.Samples.Common</Import_RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)Commands\BaseSettings.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Commands\DefaultCommand.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\VerbosityLevel.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\ConsoleDisplayWriter.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\IDisplayWriter.cs" />
+  </ItemGroup>
+</Project>
